Better naming `HasTemplateItems` to `HasTemplateItem`.
